pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "directed-inputs-class"
dynamic = ["version"]
description = "Directed inputs class consumes and processes inputs from sources beyond args and kwargs"
requires-python = ">=3.8"
license = "MIT"
readme = "README.md"
keywords = ["python3"]
authors = [{ name = "Jon Bogaty", email = "jon@jonbogaty.com" }]
maintainers = [{ name = "Jon Bogaty", email = "jon@jonbogaty.com" }]
classifiers = [
    "Development Status :: 5 - Production/Stable",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3.8",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Typing :: Typed",
]
dependencies = [
    "case-insensitive-dictionary>=0.2.1",
    "deepmerge>=1.1.1",
    "extended-data-types>=1.0.2",
    "future>=1.0.0",
    "pydantic>=2.8.2",
    "pydantic-settings>=2.4.0",
]

[project.urls]
Documentation = "https://github.com/jbcom/directed-inputs-class#readme"
Issues = "https://github.com/jbcom/directed-inputs-class/issues"
Source = "https://github.com/jbcom/directed-inputs-class"

[project.optional-dependencies]
tests = [
    "pytest>=8.2.2",
    "pytest-cov>=5.0.0",
    "pytest-mock>=3.14.0",
    "pytest-xdist>=3.6.1",
    "coverage[toml]>=7.6.0",
]

typing = [
    "mypy>=1.0.0",
    "types-PyYAML>=6.0.12.20240724",
    "sortedcontainers-stubs>=2.4.2",
]

docs = [
    "sphinx>=7.2,<7.5",
    "sphinxawesome-theme>=5.2.0",
    "sphinx-autodoc2>=0.5.0",
    "myst-parser>=3.0.1",
    "docutils>=0.17",
]

[tool.hatch.version]
path = "src/directed_inputs_class/__init__.py"

[tool.hatch.envs.default]
type = "virtual"
path = ".venv"

[tool.pytest.ini_options]
addopts = ["-ra", "--strict-markers", "--strict-config"]
testpaths = "tests"
xfail_strict = true
filterwarnings = []

[tool.coverage.run]
branch = true
parallel = true
source = ["directed_inputs_class"]

[tool.coverage.paths]
source = ["src", ".tox/py*/**/site-packages"]

[tool.coverage.report]
show_missing = true
skip_covered = true
exclude_lines = ["no cov", "if __name__ == .__main__.:", "if TYPE_CHECKING:"]

[tool.codespell]
skip = '*.js,*.css,*.inv,*.lock,*.toml,*.json,*.yml,*.yaml'
count = true

[tool.interrogate]
omit-covered-files = true
verbose = 2
fail-under = 100
whitelist-regex = ["test_.*"]

[tool.black]
line-length = 88

[tool.ruff]
src = ["directed_inputs_class", "tests"]
exclude = ["docs"]

[tool.ruff.lint]
select = ["ALL"]
ignore = ["ANN", "COM", "E501", "FBT", "PLR0913", "C901"]

[tool.ruff.lint.pydocstyle]
convention = "google"

[tool.ruff.lint.per-file-ignores]
"tests/*" = ["ANN", "D", "E", "S"]

[tool.ruff.lint.isort]
lines-between-types = 1
lines-after-imports = 2

[tool.mypy]
strict = true
pretty = true

[[tool.mypy.overrides]]
module = "wrapt.*,"
ignore_missing_imports = true

[tool.commitizen]
name = "cz_conventional_commits"
version = "0.1.0"
version_files = [
    "src/directed_inputs_class/__init__.py:__version__",
]
tag_format = "$version"
annotated_tag = true
update_changelog_on_bump = true
bump_message = "New release: $current_version → $new_version [skip actions]"
template = "docs/templates/CHANGELOG.md.j2"
